No more "byte" Use "int" instead

+++ src/model/constants/Goal.java
@@ -1,14 +1,14 @@
 package model.constants;
 
 public class Goal {
-	private byte goal;
+	private int goal;
 
-	public Goal(byte goal) {
+	public Goal(int goal) {
 		super();
 		this.goal = goal;
 	}
 
-	public byte getGoal() {
+	public int getGoal() {
 		return goal;
 	}
 
@@ -17,12 +17,12 @@ public class Goal {
 	}
 	
 	//The static Goals
-	public static final Goal ForFun = new Goal((byte) 1);
-	public static final Goal ToMeet = new Goal((byte) 2);
-	public static final Goal GetDate = new Goal((byte) 3);
-	public static final Goal SeriousRelation = new Goal((byte) 4);
-	public static final Goal DidIt = new Goal((byte) 5);
-	public static final Goal OTHER = new Goal((byte) 6);
+	public static final Goal ForFun = new Goal((int) 1);
+	public static final Goal ToMeet = new Goal((int) 2);
+	public static final Goal GetDate = new Goal((int) 3);
+	public static final Goal SeriousRelation = new Goal((int) 4);
+	public static final Goal DidIt = new Goal((int) 5);
+	public static final Goal OTHER = new Goal((int) 6);
 	
 	
 }
